Bug 15568: Fix display default circ rules
To reproduce:
Go on the admin circ rules page (admin/smart-rules.pl), switch the
dropdown list from specific to "all libraries".
The rules are not displayed anymore.

Test plan:
Apply this patch and confirm that the rules are now displayed correctly.
Make sure the enh introduced by bug 11625 is not broken.

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>

Signed-off-by: Jesse Weaver <jweaver@bywatersolutions.com>