Bug 15569: Do not displayed "automatic renewal" if the patron cannot checkout
If the patron cannot checkout (debarred) and on-site checkout is
enabled, the "Automatic renewal" checkbox should not be displayed on the
circulation screen.

Test plan:
Enable OnSiteCheckouts and OnSiteCheckoutsForce
Debar a patron and go on the circulation page.
Without this patch, the automatic renewal checkbox is displayed.
With this patch applied, it's not.

Patch works as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com