Bug 15652: Allow current date in datepicker on opac-reserve
If you enable AllowHoldDateInFuture (and its OPAC variant), you get
an additional "Hold starts on date". In that field and in "Hold not
needed after" it was not possible to enter the current date.

Although it does not really make any difference if you leave the startdate
empty or put the current date in it, we concluded that it still makes
a psychological difference. Some patrons were confused when placing a
hold because the first date to select was tomorrow. Instead of telling
them "Yes, but you can leave it empty", we should just allow the current
date as well.

The expiration date ("not needed after") also started on tomorrow. (This
field can really be empty.) But there is actually no reason not to allow
today here too. If the patron only wants it today, why force tomorrow?

How is this arranged? Well, this is one of the simplest Koha patches ever
written. It only changes the minDate option of both date pickers in one
line from 1 to 0.

Test plan:
[1] Allow future hold dates via systempreferences.
[2] Verify that you can enter today in both fields.
[3] Put today in start and expiration date. Place the hold and verify that
    you can confirm the hold, check out and check in again.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com
(cherry picked from commit 904a358da38e2945125170c3e9896193f0c288ce)
Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>
(cherry picked from commit ae2c66c601a38a30d7ed8e8a074dbb0282de6057)
Signed-off-by: Frédéric Demians <f.demians@tamil.fr>