Bug 15784: Library deletion warning is incorrectly styled
When deleting a library in Administration ->  Libraries and groups, the
deletion confirmation message is not styled with the standard "dialog
alert" <div>. This patch modifies the confirmation markup to match the
standard.

To test, apply the patch and go to Administration -> Libraries and
groups.

- Try deleting a library. The confirmation box should have the standard
  "dialog alert" class.
- Test the 'confirm' and 'cancel' actions and make sure both complete
  correctly.

Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>
(cherry picked from commit cfb46e45cb74fdc4a42bb48df78fe820d31fdc86)
Signed-off-by: Frédéric Demians <f.demians@tamil.fr>