Bug 15742: Remove unnecessary loop in j2a cronjob
If no branchcode is given, all the libraries are retrieved and the same
query (so without using the libraries loop) is executed for each
library.

Test plan:
Use the j2a.pl cronjob to change the category of a child patron
If a branchcode is passed to the script, only the children from this
branchcode should be updated.
But if it is not passed, all children of the DB should be updated.

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com
(cherry picked from commit cf047247893b04de820ff51e9b2d8f43773af332)
Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>
(cherry picked from commit 93c5535eb5498887052d5c44d661b99b3f550461)
Signed-off-by: Frédéric Demians <f.demians@tamil.fr>