Bug 15790 - Don't delete a MARC framework if existing records use that framework
If a MARC framework is in use by any records it should not be possible
to delete it. Deleting a framework which is in use will leave the
deleted framework code in those records.

This patch modifies the deletion confirmation dialog so that it will not
let the user confirm the deletion if there are records using it.

To test, apply the patch and go to Administration -> MARC frameworks

1. Try to delete a framework which is in use by 1 or more records. The
   resulting confirmation dialog should prevent you from proceeding.
2. Try to delete a framework which is not in use. Test both canceling
   and confirming that deletion to make both function correctly.

Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>
(cherry picked from commit 18820b6c8fc30fb25c99e0e0710d08eda2bc745c)
Signed-off-by: Frédéric Demians <f.demians@tamil.fr>