Bug 15844 - Correct JSHint errors in staff-global.js
This patch corrects minor errors in staff-global.js according to
warnings raised by JSHint.

To test, apply the patch and run JSHint on staff-global.js. No warnings
should be raised.

To confirm that the behavior of the JS is not changed, clear your
browser cache if necessary and:

- Confirm that the "Help" link opens the help window.
- Confirm that audio alerts work with local and remote sound files
- Confirm that the format() method is still working correctly by, for
  instance, opening the detail page for a bibliographic record with
  multiple items and choosing 'Edit -> Delete record' from the menu. The
  error message should read like "2 item(s) are attached to this
  record..."

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as described. No errors with JSHint CLI (6 errors fixed)

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com