Bug 15847: Correct JSHint errors in basket.js in the staff client
This patch corrects minor errors in the staff client's basket.js
according to warnings raised by JSHint.

To test, apply the patch and run JSHint on basket.js. No warnings
should be raised.

To confirm that the behavior of the JS is not changed, clear your
browser cache if necessary and:

- Test adding multiple titles to the Cart from staff client search
  results.
- Confirm that adding titles to the Cart triggers a message by the Cart
  link.
- Test clicking the "Cart" link in the header menu when the Cart popup
  is open but behind the main window.
- Testing emptying the Cart.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as described. No warnigns raised with jshint tool

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com