Bug 15071 - In OPAC search results, "checked out" status should be more visible
The prog template in the OPAC designated a color for the "unavailable"
class associated with items in search results which are checked out.
This class never made it into the Bootstrap theme. This patch adds it to
the Bootstrap theme's LESS file.

To test, apply the patch and search the OPAC for terms that will return
checked out items. In the list of search results, on the "availability"
line, the "Checked out" text should now be red. The change is visible
with XSLT on or off.

Signed-off-by: Nicole C Engard <nengard@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
(cherry picked from commit 4270c1443f1a66de2c9ba7f781176c8c3914c88b)
Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>