Bug 15922: Show authorized value description in staff client search results for lost, withdrawn, and damaged
If an item in the staff client search results list is lost, withdrawn,
or damaged, the actual authorized value description for each of those
statuses should be displayed rather than the generic term. This patch
adds use of the AuthorisedValues template plugin to output the correct
description.

To test, apply the patch and perform a catalog search in the staff
client which will return results with various values for lost, damaged,
and withdrawn.

Confirm that the correct lost, damaged, or withdrawn status appears for
each item rather than the generic term.

Followed test plan, statuses appear as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>